# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SQL Practice API
  version: "1.0.0"
  description: |
    API for the SQL Practice app (interactive SQL tasks checked against a real
    PostgreSQL database, sessions/learning plans, login/role system).

    **This spec is hand-maintained, not generated from code.** It documents a
    first, deliberately partial slice of the API (see "Phase 1" scope below) —
    if anything here looks stale compared to the actual route code, trust the
    code and update this file.

    ### Authentication
    This API uses a session cookie (`connect.sid`, set by `express-session` +
    `connect-pg-simple`), **not** bearer tokens. To use "Try it out" below:
    1. Call `POST /api/auth/login` with a real username/password — a valid
       session cookie is stored by your browser automatically.
    2. Every subsequent "Try it out" call from this same page reuses that
       cookie automatically — no token to copy/paste anywhere.
    3. Call `POST /api/auth/logout` to end the session.

    This only works because Swagger UI is served by the backend itself
    (same origin as `/api/*`) — see `src/routes/docs.js`.

    ### Phase 1 scope
    Only a first, "safe to explore" slice of the API is documented here:
    health, auth, datasets, tasks (read + solution), the query sandbox, and
    the table browser. Endpoints with heavier role/ownership matrices
    (sessions, users, progress, mentor assignments, mentor students) are
...
    full API surface in the meantime.
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: >-
        Session cookie set by POST /api/auth/login. httpOnly, sameSite=lax,
        secure in production. There is no bearer-token alternative.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
      required: [error]
      example:
        error: "Not authenticated."
...
    Session:
      type: object
      description: >-
        A learning plan row, enriched with read-only dataset + ownership
        metadata (owner_username/owner_role, created_by_username,
        archived_by_username — the latter two are null if never set, or if
        that user was since deleted).
      properties:
        id: { type: integer }
        user_id: { type: integer, description: "The session's owner." }
        created_by_user_id: { type: integer, nullable: true }
        name: { type: string }
        description: { type: string, nullable: true }
        plan_type: { type: string, enum: [topic, category, project] }
        status:
          type: string
          enum: [active, completed]
          description: Completion state — orthogonal to archived_at (see below).
        dataset_id: { type: integer, nullable: true }
        archived_at:
          type: string
          format: date-time
          nullable: true
          description: null = visible/normal session. Set = archived (hidden from the default list).
        archived_by_user_id: { type: integer, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        last_opened_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        dataset_key: { type: string, nullable: true }
        dataset_name: { type: string, nullable: true }
        schema_name: { type: string, nullable: true }
        dataset_type: { type: string, nullable: true }
        owner_username: { type: string, nullable: true }
        owner_role: { type: string, enum: [admin, mentor, student], nullable: true }
        created_by_username: { type: string, nullable: true }
        archived_by_username: { type: string, nullable: true }
      example:
        id: 42
        user_id: 7
        created_by_user_id: 7
        name: "JOINs practice"
        description: null
        plan_type: topic
        status: active
        dataset_id: 1
        archived_at: null
        archived_by_user_id: null
        completed_at: null
        last_opened_at: "2026-07-01T10:00:00.000Z"
        created_at: "2026-06-20T08:00:00.000Z"
        dataset_key: academic
        dataset_name: Academic
        schema_name: academic
        dataset_type: official
        owner_username: jane_student
        owner_role: student
        created_by_username: jane_student
        archived_by_username: null
...
    MentorStudentSummary:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        role: { type: string, enum: [student] }
        assigned_at: { type: string, format: date-time }
        active_sessions: { type: integer }
        completed_sessions: { type: integer }
        archived_sessions: { type: integer }
        solved_count: { type: integer }
        last_activity:
          type: string
          format: date-time
          nullable: true
          description: MAX(last_attempt_at) across ALL of the student's sessions (Run Query and Check Answer both count).
    MentorStudentSession:
      description: >-
        A Session, plus a lightweight solved/attempted count for this
        specific session (not a plan-scoped fraction — see the route
        description for why).
      allOf:
        - $ref: "#/components/schemas/Session"
        - type: object
          properties:
            solved_count: { type: integer }
            attempted_count: { type: integer }
...
security:
  - sessionCookie: []
paths:
...
  /api/mentor/students:
    get:
      tags: [Mentor Students]
      summary: List the caller's own assigned students
      description: >-
        **Mentor-only — deliberately excludes admin too** (an admin's own
        "assigned roster" isn't a meaningful concept; admins already have
        GET /api/mentor-assignments for full oversight). Always scoped to
        `req.actingUser` — there is no targetUserId here, unlike most other
        review-mode routes.
      operationId: listMentorOwnStudents
      x-required-role: mentor
      x-visibility: own
      responses:
        "200":
          description: The caller's assigned students.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer }
                    username: { type: string }
                    role: { type: string, enum: [student] }
                    assigned_at: { type: string, format: date-time }
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Authenticated but not a mentor (including admin).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/mentor/students/summary:
    get:
      tags: [Mentor Students]
      summary: Aggregated per-student stats for the caller's own roster
      description: >-
        Mentor-only, same scope as GET /api/mentor/students (not admin). One
        query for the whole roster (session counts, solved count, last
        activity) instead of N-per-student round trips.
      operationId: getMentorStudentsSummary
      x-required-role: mentor
      x-visibility: own
      responses:
        "200":
          description: Per-student aggregated stats.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MentorStudentSummary" }
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Authenticated but not a mentor (including admin).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/mentor/students/{studentId}/sessions:
    get:
      tags: [Mentor Students]
      summary: Full session history (including archived) for one student
      description: >-
        Mentor (only if assigned, via canAccessStudent) or admin (always) —
        the path param is never trusted alone, same re-authorization pattern
        as every other targetUserId-style route. **Students are explicitly
        blocked even for their own id** — this route is meant for
        mentor/admin oversight, not self-viewing (use GET /api/sessions for
        that). Unlike GET /api/sessions, this always includes archived
        sessions — the point is a complete history, not "what to resume."
        Each session carries a lightweight solved/attempted count (not a
        plan-scoped fraction).
      operationId: getMentorStudentSessions
      x-visibility: assigned-student, admin-any
      parameters:
        - name: studentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: The student's full session history, most recent first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MentorStudentSession" }
        "400":
          description: Invalid student id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: "Caller is a student (even for their own id), or a mentor not assigned to this student."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }